requirements.txt
boto3==1.12.39
cached_property
cachetools
click
cryptography
dataclasses
django-cachalot
django-cache-memoize
django-cors-headers
django-environ
django-extensions
django-localflavor==3.0.1
django-markdownx
django-naomi
django-recaptcha
django-redis
django-rq==2.3.1
django-storages==1.9.1
django>=2.0
djangorestframework
git+https://github.com/turicas/rows.git@develop#egg=rows
git+https://github.com/turicas/rows.git@develop#egg=rows[ods]
gunicorn
ipython
jinja2
lxml==4.5.0
markdown
networkx==2.1
openpyxl
psycopg2-binary
py2neo==3.1.2
pytest-django==3.2.1
pytz
requests
requests-cache
sendgrid-django
six>=1.13
tqdm
whitenoise
xlrd
xlwt
